+++ tests/database/factories/users/UsersFactory.php
@@ -2,10 +2,13 @@
 
 use Vegvisir\TrustNoSql\Tests\Infrastructure\Models\User as Model;
 
-Model::where(1)->delete();
+if (null == Cache::get('create_users')) {
+    Model::where(1)->delete();
 
-$users = test_seed_users();
+    $users = test_seed_users();
 
-foreach ($users as $user) {
-    Model::create($user);
+    foreach ($users as $user) {
+        Model::create($user);
+    }
+    Cache::put('create_users', 2);
 }
